+++ tests/test_end_to_end.py
@@ -110,7 +110,15 @@ class TestEndToEnd(unittest.TestCase):
 
     def test_no_model(self):
         tbmodel = torchbearer.Trial(None)
+
+        import warnings
+        with warnings.catch_warnings(record=True) as w:
+            tbmodel.run()
+            self.assertTrue(len(w) == 1)
+
+        self.assertTrue(torchbearer.trial.MockModel()(torch.rand(1)) is None)
+
+    def test_no_train_steps(self):
+        tbmodel = torchbearer.Trial(None)
+        tbmodel.for_val_steps(10)
         tbmodel.run()
-
-        self.assertTrue(tbmodel.state[torchbearer.MODEL](torch.rand(1)) is None)
-